Add Network permission change earthquake depth type from int to double

+++ app/src/main/java/com/github/liangyu/earthquakes/data/response/Earthquake.kt
@@ -2,7 +2,7 @@ package com.github.liangyu.earthquakes.data.response
 
 data class Earthquake(
     val datetime: DateTime,
-    val depth: Int,
+    val depth: Double,
     val eqid: String,
     val lat: Double,
     val lng: Double,
